backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
Pillow>=10.0
opencv-python-headless>=4.8
numpy>=1.24
python-dotenv>=1.0
gunicorn>=21.2
